ensure correct order of evaluation for binary operations
The order of evaluation for binary operations was always from right to left.
This is bad as there are implicit rules the user would expect.
This means that we have to ensure that some operations happen before others.
That is now happening thanks to some intermediate chain logic as can be found in
the Binary Parser.