prevent optimization for hashvalues
Logic as defined by translators in the L20n DSL might break in case we optimize this.
This would for example happen when we have a property expression such as 'foo.bar.baz'.
In case we optimize 'bar' away, because 'baz' is its only value, that expression wouldn't be valid anymore.